5 - вынес стили в отдельный файл для компонента Game и его фрагмента GameTabs
это сделано для лучшей читаемости кода
