+++ src/Components/Game/GameTabs.js
@@ -4,15 +4,8 @@ import Tabs from '@material-ui/core/Tabs';
 import Tab from '@material-ui/core/Tab';
 import { withStyles } from '@material-ui/core/styles';
 import PropTypes from 'prop-types';
+import styles from "./GameStyles";
 
-const styles = () => ({
-    tabs: {
-        flexGrow: 1,
-    },
-    tab: {
-        minHeight: 30,
-    },
-});
 
 function GameTabs(props) {
     const { classes, tabValue, handleChangeTab } = props;
